Rename files to use UpperCamelCaseCluster/Logic for ota-provider
